+++ samples/JitterOpenGLDemo/JitterOpenGLDemo/DrawStuffOtk.cs
@@ -46,14 +46,14 @@ namespace JitterOpenGLDemo
     {
         public static float[] ToFloat(JVector vector)
         {
-            return new float[4] { vector.X, vector.Y, vector.Z, 0.0f };
+            return new float[4] { (float)vector.X, (float)vector.Y, (float)vector.Z, 0.0f };
         }
 
         public static float[] ToFloat(JMatrix matrix)
         {
-            return new float[12] { matrix.M11, matrix.M21, matrix.M31, 0.0f,
-                                   matrix.M12, matrix.M22, matrix.M32, 0.0f,
-                                   matrix.M13, matrix.M23, matrix.M33, 1.0f };
+            return new float[12] { (float)matrix.M11, (float)matrix.M21, (float)matrix.M31, 0.0f,
+                                   (float)matrix.M12, (float)matrix.M22, (float)matrix.M32, 0.0f,
+                                   (float)matrix.M13, (float)matrix.M23, (float)matrix.M33, 1.0f };
         }
     }
